requirements.txt
flask
mysql-connector-python
apscheduler
python-dotenv
